+++ internal/stack/actions.go
@@ -312,15 +312,28 @@ var Swap = newAction(
 	"pop 'a', 'b'; push 'b', 'a'",
 )
 
-var Roll = newAction(
-    func(so *StackOperator) (string, error) {
-        fS := make([]float64, len(so.Stack.Values))
+var Froll = newAction(
+	func(so *StackOperator) (string, error) {
+		newVals := make([]float64, len(so.Stack.Values))
+        newVals[0] = so.Stack.Values[len(so.Stack.Values)-1]
         for i, f := range so.Stack.Values[:len(so.Stack.Values)-1] {
-            fS[i+1] = f
+			newVals[i+1] = f
+		}
+		so.Stack.Values = newVals
+		return so.Stack.Display(), nil
+	}, 2, 2,
+	"roll the stack to the right one position",
+)
+
+var Rroll = newAction(
+	func(so *StackOperator) (string, error) {
+		newVals:= make([]float64, len(so.Stack.Values))
+        for i, f := range so.Stack.Values[1:] {
+            newVals[i] = f
         }
-        fS[0] = so.Stack.Values[len(so.Stack.Values)-1]
-        so.Stack.Values = fS
-        return so.Stack.Display(), nil
-    }, 2, 2,
-    "roll the stack to the right one position",
+        newVals[len(newVals)-1] = so.Stack.Values[0]
+        so.Stack.Values = newVals
+		return so.Stack.Display(), nil
+	}, 2, 2,
+	"roll the stack to the left one position",
 )